requirements.txt
pytest==6.2.5
Appium-Python-Client==2.7.1
selenium==4.7.2
python_dotenv==1.0.1
